Fix performance regression with avx enabled XMMatrixMultiply[Transpose]
XMMatrixMultiply and XMMatrixMultiplyTranspose were performing
worse than the sse2 counterpart. It seems that pointer `add` with
a constant value may not be optimized in some cases.

A new `idx!` variant is now used that transmutes the pointer as
as a slice and indexes without pointer addition.

Although XMMatrixDecompose also makes use of pointer `add`, the slice
transmute didn't produce any noticeable speedup and so XMMatrixDecompose
has been left unmodified.